+++ src/Percolation.hpp
@@ -1,7 +1,9 @@
 #pragma once
 #include "UnionFind.hpp"
+
 #include <memory>
 #include <vector>
+#include <stdint.h>
 
 
 //class UnionFind;
@@ -9,16 +11,16 @@
 class Percolation {
   public:
     // create N-by-N grid, with all sites blocked
-    Percolation(size_t N);
+    Percolation(uint64_t N);
     
     // open site (row i, column j) if it is not open already
-    void open(size_t i, size_t j);
+    void open(uint64_t i, uint64_t j);
     
     // is site (row i, column j) open?
-    bool isOpen(size_t i, size_t j);
+    bool isOpen(uint64_t i, uint64_t j);
     
     // is site (row i, column j) full?
-    bool isFull(size_t i, size_t j) {};
+    bool isFull(uint64_t i, uint64_t j) {};
     
     // does the system percolate?
     bool percolates();
@@ -27,9 +29,9 @@ class Percolation {
     void display();
     
   private:
-    size_t N_;
+    uint64_t N_;
     std::unique_ptr<UnionFind> uf_;
     std::vector<bool> grid_;
     
-    size_t flatten(size_t i, size_t j);
+    uint64_t flatten(uint64_t i, uint64_t j);
 };
